third_party/WebKit/LayoutTests/css3/motion-path/use-count-offset-position.html

<!DOCTYPE html>
<head>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
  <style>
  .n { offset-position: auto; }
  </style>
</head>

<div id="target"></div>
<script>
var CSSOffsetInEffect = 1617; // From UseCounter.h

test(function() {
  assert_false(internals.isUseCounted(document, CSSOffsetInEffect));

  var styleElement = document.createElement('style');
  styleElement.textContent = ".p { offset-position: center; }";
  document.head.appendChild(styleElement);

  assert_true(internals.isUseCounted(document, CSSOffsetInEffect));
});
</script>
